Norme | Clean | -l /

+++ ft_maxlen.c
@@ -6,7 +6,7 @@
 /*   By: lscariot <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/02/10 15:40:37 by lscariot          #+#    #+#             */
-/*   Updated: 2016/02/11 09:16:54 by lscariot         ###   ########.fr       */
+/*   Updated: 2016/02/13 00:55:52 by lscariot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,21 +31,21 @@ t_lenmax	ft_maxlen(t_files *files)
 	ft_declare(&max);
 	while (files != NULL)
 	{
-		if (ft_strlen(files->name) > max.name)
+		if ((int)ft_strlen(files->name) > max.name)
 			max.name = ft_strlen(files->name);
-		if (ft_strlen(files->links) > max.links)
+		if ((int)ft_strlen(files->links) > max.links)
 			max.links = (ft_strlen(files->links) + 1);
-		if (ft_strlen(files->user) > max.user)
+		if ((int)ft_strlen(files->user) > max.user)
 			max.user = ft_strlen(files->user);
-		if (ft_strlen(files->group) > max.group)
+		if ((int)ft_strlen(files->group) > max.group)
 			max.group = (ft_strlen(files->group) + 1);
-		if (ft_strlen(files->size) > max.size)
+		if ((int)ft_strlen(files->size) > max.size)
 			max.size = (ft_strlen(files->size) + 1);
-		if (ft_strlen(files->month) > max.month)
+		if ((int)ft_strlen(files->month) > max.month)
 			max.month = ft_strlen(files->month);
-		if (ft_strlen(files->number) > max.number)
+		if ((int)ft_strlen(files->number) > max.number)
 			max.number = ft_strlen(files->number);
-		if (ft_strlen(files->hour) > max.hour)
+		if ((int)ft_strlen(files->hour) > max.hour)
 			max.hour = ft_strlen(files->hour);
 		files = files->next;
 	}
